backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
<<<<<<< Updated upstream
sqlparse==0.5.5
=======
sqlparse==0.5.5
>>>>>>> Stashed changes
